Added Menno's idea for catching KeyboardInterrupt.  This allows the program to do things like hit the next retry, jump to the next mirror, etc. when the user hits ctrl-c.
Also modified the behavior of failure_callback slightly.  It now gets
called for EVERY failure, even the last one.  That is, if there's a
failure on the last try, the failure callback gets called and
(assuming it doesn't raise anything) the exception gets raised.
Previously, it would not be called on the last failure.
This is another MINOR compatibility break.