handle endless-data problems safely:
"A malicious server could cause libcurl
to download an infinite amount of data, potentially causing all of memory or
disk to be filled.  Setting the CURLOPT_MAXFILESIZE_LARGE option is not
sufficient to guard against this.  Instead, the app should monitor the
amount of data received within the write or progress callback and abort once
the limit is reached."

had to restructure a good bit of the error handling to do this but it works for both
endless headers and endless content.