Patch to allow url arguments to be of type unicode
Description of problem:

I am working on some code that handed a url as a unicode string to urlgrabber.
This lead to a problem with two sections of code.  One is in mirror.py where
we're only allowing byte strings.  The other is in grabber.py which hands off
the unicode strings to pycurl.  pycurl only understands byte strings and raises
an error if a unicode string is given to it.

We need to transform any unicode strings into byte strings before it gets to
curl.  Since we're doing that, we might as well allow unicode strings in
mirror.py and transform those to byte strings as well.