Sort mirrors using TH statistics.
The main use case for this is when mirror #1 is down.  Parallel
downloader avoids such mirror, but legacy mirror group code does not.
This patch brings these in line.

Legacy users that don't set the 'timedhosts' option are not affected
(all mirrors evaluate to default speed and sorting is stable)

caveats: This runs after the fastestmirror plugin (if enabled) has
sorted the mirror list.  When TH file is used and contains enough
host stats, FM order is mostly ignored.