Add b'' markers in various places
This is in preparation for python3 compatiblity, but not enough to
reach it. Under python2, this has no effect, because b'' is fully
ignored. b'' was supported under python2.6, so this should not raise
backwards-compatibility issues.

I'm making this a separate commit because it touches a lot of lines
but is otherwise trivial.