Simplify mirror conversion to utf8
The code would (potentially) skip some items in the mirror list if
they were not strings. But this would be a programming error in the
caller. Let's assume that mirror list does not contain garbage with
unexpected types and do the conversion unconditionally.

Note: mirror URLs are still stored encoded, because that seems to work
better with urllib.