Add explicit encode/decode calls
This is the patch that has the most potential for trouble in the whole
series, because it affects python2 behaviour directly.

io.BytesIO is the same as StringIO under python2, so this should have
no effect on python2. Under python3 it is necessary to allow reading
bytes from a byte data source.

Under python2, encoding of an already encoding string is allowed, and
actually works fine (is idempotent) for ASCII strings. So the effect
of the .encode() calls under python2 should be limited. Under python3,
they are of course necessary and can only be done once. So if there
are errors here, they should show up when running under python3 pretty
easily.