Make the 'broker' mirror really broken
I guess those files are not supposed to be there, because otherwise
the test don't get the exception they expect. The test is updated
to expect 404, not 403, because the file is simply missing. It would
be possible to change the server configuration to return 403, but
this doesn't seem particularly important for this test. So let's just
ask for a file that doesn't exist and expect 404.