Fix the confused license header in urlgrabber/__init__.py file
urlgrabber was relicensed to LGPLv2+ a long time ago, and this file
seemed to have only a partially updated license header for this.

It references the GNU Library General Public License, so clearly
the intent was the header was supposed to be fully updated for it,
so let's just do that by switching to the header used in the other
source files.